Update and rename lab1_couse1.py to lab1_course1.py
